2018 round 2 submissions
This problem A code is buggy wrt chat I had in mind, but it happened
to pass the small test case on first try.  So once I noticed the bug,
I had to spend quite some time pondering whether or not to resubmit.
(new rules, yay!)  I concluded the bug did issue spurious ramps, but
only in places where they're useless, so the code would remain.

It passed large as well.  Whew!